fbla_ebusiness_2023/backend/requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
numpy
pandas
argon2-cffi
